edit

+++ src/main/java/web/dao/UserDao.java
@@ -6,12 +6,12 @@ import web.model.User;
 public interface UserDao {
     List<User> findAll();
 
-    User findById(Integer id);
+    User findById(long id);
 
     void save(User user);
 
     void update(User user);
 
-    void delete(int id);
+    void delete(long id);
 
 }
