+++ src/main/java/web/model/User.java
@@ -12,13 +12,14 @@ import javax.persistence.Table;
 public class User {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Integer id;
+    private long id;
     @Column(name = "name")
     private String name;
     @Column(name = "last_name")
     private String lastName;
     @Column(name = "is_person_studying_java")
     private boolean PersonWhoStudiesJava;
+
 
 
 
@@ -31,11 +32,11 @@ public class User {
         PersonWhoStudiesJava = personWhoStudiesJava;
     }
 
-    public Integer getId() {
+    public long getId() {
         return id;
     }
 
-    public void setId(Integer id) {
+    public void setId(long id) {
         this.id = id;
     }
